pyproject.toml
[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta:__legacy__"

[project]
name = "org-ops-tools"
version = "1.0.0"
description = "CLI tools for ops monitoring -- SQLMI, Azure DevOps, shared auth"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "packaging>=23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
    "pytest-mock",
    "ruff>=0.15",
    "mypy>=1.19",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
